# docs/swagger.yaml
definitions:
...
  feed.ChangeFeedAllowListReq:
    properties:
      air_conditioner:
        type: boolean
      grade:
        type: boolean
      holiday:
        type: boolean
      light:
        type: boolean
      muxi:
        type: boolean
    type: object
  feed.ClearFeedEventReq:
    properties:
      feed_id:
        type: integer
    type: object
  feed.FeedEvent:
    properties:
      content:
        type: string
      created_at:
        type: integer
      extend_fields:
        additionalProperties:
          type: string
        type: object
      id:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  feed.GetFeedAllowListResp:
    properties:
      air_conditioner:
        type: boolean
      grade:
        type: boolean
      holiday:
        type: boolean
      light:
        type: boolean
      muxi:
        type: boolean
    type: object
  feed.GetFeedEventsResp:
    properties:
      read_events:
        items:
          $ref: '#/definitions/feed.FeedEvent'
        type: array
      unread_events:
        items:
          $ref: '#/definitions/feed.FeedEvent'
        type: array
    type: object
  feed.GetToBePublicMuxiOfficialMSGResp:
    properties:
      msg_list:
        items:
          $ref: '#/definitions/feed.MuxiOfficialMSG'
        type: array
    type: object
  feed.MuxiOfficialMSG:
    properties:
      content:
        type: string
      extend_fields:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      public_time:
        type: integer
      title:
        type: string
    type: object
  feed.PublicMuxiOfficialMSGReq:
    properties:
      content:
        type: string
      extend_fields:
        additionalProperties:
          type: string
        type: object
      later_time:
        type: integer
      title:
        type: string
    type: object
  feed.PublicMuxiOfficialMSGResp:
    properties:
      content:
        type: string
      extend_fields:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      public_time:
        type: string
      title:
        type: string
    type: object
  feed.ReadFeedEventReq:
    properties:
      feed_id:
        type: integer
    type: object
  feed.RemoveFeedTokenReq:
    properties:
      token:
        type: string
    type: object
  feed.SaveFeedTokenReq:
    properties:
      token:
        type: string
    type: object
  feed.StopMuxiOfficialMSGReq:
    properties:
      id:
        type: string
    type: object
...
  web.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /feed/changeFeedAllowList:
    post:
      consumes:
      - application/json
      description: 修改已登录用户的feed订阅白名单设置
      parameters:
      - description: 白名单设置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feed.ChangeFeedAllowListReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - BearerAuth: []
      summary: 修改feed订阅白名单
      tags:
      - feed订阅
  /feed/clearFeedEvent:
    post:
      consumes:
      - application/json
      description: 清除指定用户的feed订阅事件,feedid有三种状态0表示全部清除(不加参数默认为0),-1表示清除已读,指定feedid表示只清除这个id的feed消息
      parameters:
      - description: feed订阅事件ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feed.ClearFeedEventReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - BearerAuth: []
      summary: 清除feed订阅事件
      tags:
      - feed订阅
  /feed/getFeedAllowList:
    get:
      consumes:
      - application/json
      description: 获取已登录用户的feed订阅白名单设置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/feed.GetFeedAllowListResp'
              type: object
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - BearerAuth: []
      summary: 获取feed订阅白名单
      tags:
      - feed订阅
  /feed/getFeedEvents:
    get:
      description: 获取已登录用户的所有feed订阅事件（包括已读和未读）
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/feed.GetFeedEventsResp'
              type: object
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - Bearer Auth: []
      summary: 获取feed订阅事件
      tags:
      - feed订阅
  /feed/getToBePublicOfficialMSG:
    get:
      consumes:
      - application/json
      description: 获取计划发布的官方消息，仅限管理员操作
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/feed.GetToBePublicMuxiOfficialMSGResp'
              type: object
        "403":
          description: 没有访问权限
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - BearerAuth: []
      summary: 获取待发布的官方消息
      tags:
      - feed订阅
  /feed/publicMuxiOfficialMSG:
    post:
      consumes:
      - application/json
      description: 发布木犀官方消息，仅限管理员操作
      parameters:
      - description: 木犀官方消息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feed.PublicMuxiOfficialMSGReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/feed.PublicMuxiOfficialMSGResp'
              type: object
        "403":
          description: 没有访问权限
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - BearerAuth: []
      summary: 发布木犀官方消息
      tags:
      - feed订阅
  /feed/readFeedEvent:
    post:
      consumes:
      - application/json
      description: 标注feed订阅事件为已读
      parameters:
      - description: feed订阅事件ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feed.ReadFeedEventReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - Bearer Auth: []
      summary: 标注feed订阅事件为已读
      tags:
      - feed订阅
  /feed/removeFeedToken:
    post:
      consumes:
      - application/json
      description: 删除已登录用户的feed订阅Token
      parameters:
      - description: feed订阅Token
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feed.RemoveFeedTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - BearerAuth: []
      summary: 删除feed订阅Token
      tags:
      - feed订阅
  /feed/saveFeedToken:
    post:
      consumes:
      - application/json
      description: 保存已登录用户的feed订阅Token
      parameters:
      - description: feed订阅Token
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feed.SaveFeedTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - BearerAuth: []
      summary: 保存feed订阅Token
      tags:
      - feed订阅
  /feed/stopMuxiOfficialMSG:
    post:
      consumes:
      - application/json
      description: 停止木犀官方消息，仅限管理员操作
      parameters:
      - description: 停止消息请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feed.StopMuxiOfficialMSGReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "403":
          description: 没有访问权限
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - BearerAuth: []
      summary: 停止木犀官方消息
      tags:
      - feed订阅
...
swagger: "2.0"